Move animation related components to dedicated subfolder
