+++ src/Components/animation/DisplayedElement.js
